Fix old buggy assumptions about various buffer sizes.
Properly define maximum sizes for principal names, ACE sizes, and ACL size.
Previously, both the size of a given ACE and the number of ACEs allowed in an
ACL were artificially constrained and caused a variety of bugs.

Special thanks to Brian De Wolf for help both on and off the linux-nfs
mailing list in reporting and debugging these issues.

Signed-off-by: David M. Richter <richterd@citi.umich.edu>